Move method login_required from utils to login_service
